// client/utils/Enum.ts
export enum ENVIRONMENT_TYPE {
  DEV= 'DEV',
  LOCAL = 'LOCAL',
  PRODUCTION = 'PRODUCTION'
}

export enum BUTTON_VARIANT {
  OUTLINED ='outlined',
  CONTAINED = 'contained',
}

export enum SIZE {
  LARGE ='large',
  MEDIUM = 'medium',
  SMALL = 'small',
}

export enum TYPOGRAPHY_VARIANT {
  H1='h1',
  H2='h2',
  H3='h3',
  H4='h4',
  H5='h5',
  H6='h6',
  BODY1 = 'body1'
}

export enum LOADING_STATUS {
  NOT_YET_STARTED = 'NOT_YET_STARTED',
  LOADING = 'LOADING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}

export enum TEXT_FIELD_VARIANT {
  OUTLINED ='outlined',
  FILLED = 'filled',
  STANDARD = 'standard',
}

export enum DROPDOWN_VARIANT{
  SIMPLE_DROPDOWN = 'SIMPLE_DROPDOWN',
  MULTI_SELECT_DROPDOWN = 'MULTI_SELECT_DROPDOWN',
}

export enum MEETING_REPEAT_INTERVAL {
  DAILY = 'DAILY',
  ALTERNATE_DAYS = 'ALTERNATE_DAYS',
  WEEKLY = 'WEEKLY',
  MONTHLY = 'MONTHLY',
}